refactor(perps): extract shared PerpsActionSheet to reduce duplication
