+++ app/components/UI/Perps/components/PerpsActionSheet/index.ts
@@ -1,0 +1,5 @@
+export { default } from './PerpsActionSheet';
+export type {
+  PerpsActionSheetOption,
+  PerpsActionSheetProps,
+} from './PerpsActionSheet.types';
